Order can now have multiple columns
Order can now have multiple columns, but there is some BC changes.

Order of columns are based of call order of getNewState, so if you
manually change order of getNewState, or add custom $order, it can be
changed again by user (or if user change URL manually). This can be
fixed by calling columns in proper order on getOrder (not sure if this
is necessary as it may be intend for programmers to change order).

$order passed to data callback and data count callback is in key =>
value form, so it is easier to build order string.
$order = [
`column` => `ASC|DESC`
]